Depend on logger-application since it was removed from Ruby 2.2
The Logger::Application class was extracted from Ruby 2.2 and moved to the
logger-application gem as it is no longer maintained:
https://bugs.ruby-lang.org/issues/9860

This requires the gem version when the Logger::Application constant is
not defined.

Ideally, soap2r should be refactored to no longer depend on the
Logger::Application class since it is no longer maintained.